# Optimized vectorSelectorSingle()