--- a/storage/memoized_iterator.go
+++ b/storage/memoized_iterator.go
@@ -19,7 +19,7 @@ import (
 	"github.com/prometheus/prometheus/tsdb/chunkenc"
 )
 
-// MemoizedSeriesIterator wraps an iterator with a buffer to look-back the previous element.
+// MemoizedSeriesIterator wraps an iterator with a buffer to look back the previous element.
 type MemoizedSeriesIterator struct {
 	it    chunkenc.Iterator
 	delta int64