I'm investigating some slow queries which, according to the profile, looks spending much time in the PromQL and I'm checking if there's any low-hanging fruit. It's difficult to find something to optimise (you've all already done a great job).

I've noticed that for some queries, a significant % of time inside PromQL is spent in `vectorSelectorSingle()`, which iterates the samples using `BufferedSeriesIterator`. This buffered iterator keeps a buffer of the returned buffer but `vectorSelectorSingle()` doesn't actually need the whole buffer but just the previous value.

In this PR I'm proposing some code duplication (sorry) introducing `MemoizedSeriesIterator` which is like `BufferedSeriesIterator` but just keeps track of the previous element. Benchmarks on `MemoizedSeriesIterator.Next()` are about 2x faster than `BufferedSeriesIterator`, but impact on the whole range query is definitely smaller (see below).


## BenchmarkBufferedSeriesIterator

```
BenchmarkBufferedSeriesIterator-12    	87461358	        13.8 ns/op	101247175509.44 MB/s	       0 B/op	       0 allocs/op
BenchmarkBufferedSeriesIterator-12    	91528774	        13.1 ns/op	111380936676.88 MB/s	       0 B/op	       0 allocs/op
BenchmarkBufferedSeriesIterator-12    	86143618	        14.9 ns/op	92586167446.73 MB/s	           0 B/op	       0 allocs/op
BenchmarkBufferedSeriesIterator-12    	90630134	        13.1 ns/op	110779457851.50 MB/s	       0 B/op	       0 allocs/op
BenchmarkBufferedSeriesIterator-12    	90841784	        13.8 ns/op	105544088430.95 MB/s	       0 B/op	       0 allocs/op
```

## BenchmarkMemoizedSeriesIterator

```
BenchmarkMemoizedSeriesIterator-12    	167776471	         7.30 ns/op	367776094718.48 MB/s	       0 B/op	       0 allocs/op
BenchmarkMemoizedSeriesIterator-12    	166226320	         7.26 ns/op	366486550585.07 MB/s	       0 B/op	       0 allocs/op
BenchmarkMemoizedSeriesIterator-12    	167046273	         7.14 ns/op	374182160447.54 MB/s	       0 B/op	       0 allocs/op
BenchmarkMemoizedSeriesIterator-12    	165566178	         7.36 ns/op	359816615878.45 MB/s	       0 B/op	       0 allocs/op
BenchmarkMemoizedSeriesIterator-12    	167822944	         7.18 ns/op	373868150052.88 MB/s	       0 B/op	       0 allocs/op
```

## RangeQuery

```
name                                                                                                       old time/op    new time/op    delta
RangeQuery/expr=a_one,steps=1-12                                                                             12.6µs ± 2%    12.0µs ± 3%   -4.71%  (p=0.008 n=5+5)
RangeQuery/expr=a_one,steps=10-12                                                                            12.9µs ± 1%    12.1µs ± 1%   -5.82%  (p=0.008 n=5+5)
RangeQuery/expr=a_one,steps=100-12                                                                           19.6µs ± 1%    18.0µs ± 0%   -8.21%  (p=0.016 n=5+4)
RangeQuery/expr=a_one,steps=1000-12                                                                          68.5µs ± 2%    61.6µs ± 0%  -10.01%  (p=0.008 n=5+5)
RangeQuery/expr=a_ten,steps=1-12                                                                             50.2µs ± 0%    47.5µs ± 0%   -5.56%  (p=0.016 n=5+4)
RangeQuery/expr=a_ten,steps=10-12                                                                            54.6µs ± 1%    51.4µs ± 1%   -5.88%  (p=0.008 n=5+5)
RangeQuery/expr=a_ten,steps=100-12                                                                            120µs ± 2%     110µs ± 0%   -8.29%  (p=0.008 n=5+5)
RangeQuery/expr=a_ten,steps=1000-12                                                                           600µs ± 2%     537µs ± 1%  -10.52%  (p=0.008 n=5+5)
RangeQuery/expr=a_hundred,steps=1-12                                                                          441µs ± 0%     423µs ± 2%   -4.20%  (p=0.008 n=5+5)
RangeQuery/expr=a_hundred,steps=10-12                                                                         484µs ± 1%     459µs ± 1%   -5.04%  (p=0.008 n=5+5)
RangeQuery/expr=a_hundred,steps=100-12                                                                       1.14ms ± 1%    1.05ms ± 1%   -7.96%  (p=0.016 n=4+5)
RangeQuery/expr=a_hundred,steps=1000-12                                                                      6.13ms ± 9%    5.51ms ±18%     ~     (p=0.095 n=5+5)
RangeQuery/expr=rate(a_one[1m]),steps=1-12                                                                   17.3µs ± 0%    17.1µs ± 0%   -1.18%  (p=0.016 n=4+5)
RangeQuery/expr=rate(a_one[1m]),steps=10-12                                                                  18.4µs ± 0%    18.2µs ± 2%     ~     (p=0.151 n=5+5)
RangeQuery/expr=rate(a_one[1m]),steps=100-12                                                                 32.3µs ± 0%    31.9µs ± 0%   -1.17%  (p=0.008 n=5+5)
RangeQuery/expr=rate(a_one[1m]),steps=1000-12                                                                 149µs ± 0%     150µs ± 1%   +0.86%  (p=0.008 n=5+5)
RangeQuery/expr=rate(a_ten[1m]),steps=1-12                                                                   54.6µs ± 2%    54.0µs ± 0%     ~     (p=0.056 n=5+5)
RangeQuery/expr=rate(a_ten[1m]),steps=10-12                                                                  65.2µs ± 0%    64.9µs ± 0%   -0.41%  (p=0.016 n=5+4)
RangeQuery/expr=rate(a_ten[1m]),steps=100-12                                                                  207µs ± 2%     203µs ± 0%   -1.59%  (p=0.008 n=5+5)
RangeQuery/expr=rate(a_ten[1m]),steps=1000-12                                                                1.54ms ±17%    1.37ms ± 1%     ~     (p=0.690 n=5+5)
RangeQuery/expr=rate(a_hundred[1m]),steps=1-12                                                                458µs ± 7%     433µs ± 0%   -5.40%  (p=0.016 n=4+5)
RangeQuery/expr=rate(a_hundred[1m]),steps=10-12                                                               555µs ± 1%     543µs ± 0%   -2.27%  (p=0.016 n=5+4)
RangeQuery/expr=rate(a_hundred[1m]),steps=100-12                                                             1.97ms ± 3%    1.92ms ± 0%   -2.46%  (p=0.008 n=5+5)
RangeQuery/expr=rate(a_hundred[1m]),steps=1000-12                                                            14.8ms ±16%    13.6ms ± 1%   -7.59%  (p=0.032 n=5+5)
RangeQuery/expr=rate(a_one[1m]),steps=10000-12                                                               1.53ms ±10%    1.48ms ± 0%     ~     (p=0.095 n=5+5)
RangeQuery/expr=rate(a_ten[1m]),steps=10000-12                                                               14.7ms ± 1%    14.6ms ± 1%     ~     (p=0.421 n=5+5)
RangeQuery/expr=rate(a_hundred[1m]),steps=10000-12                                                            146ms ± 1%     147ms ± 2%     ~     (p=1.000 n=5+5)
RangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1-12                                                  766µs ± 1%     765µs ± 1%     ~     (p=0.421 n=5+5)
RangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=10-12                                                1.25ms ± 0%    1.25ms ± 0%     ~     (p=1.000 n=4+5)
RangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=100-12                                               6.18ms ± 3%    6.12ms ± 1%     ~     (p=0.310 n=5+5)
RangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1000-12                                              55.9ms ± 3%    55.5ms ± 1%     ~     (p=0.841 n=5+5)
RangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1-12                                                 6.32ms ± 5%    6.13ms ± 0%   -3.00%  (p=0.008 n=5+5)
RangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=10-12                                                11.8ms ± 5%    11.0ms ± 1%   -6.23%  (p=0.008 n=5+5)
RangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=100-12                                               62.6ms ±10%    60.2ms ± 0%   -3.75%  (p=0.016 n=5+5)
RangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1000-12                                               545ms ± 1%     545ms ± 1%     ~     (p=0.690 n=5+5)
RangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1-12                                             61.0ms ± 0%    61.1ms ± 1%     ~     (p=1.000 n=5+5)
RangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=10-12                                             111ms ± 2%     109ms ± 1%     ~     (p=0.056 n=5+5)
RangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=100-12                                            598ms ± 2%     594ms ± 1%     ~     (p=0.690 n=5+5)
RangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1000-12                                           5.44s ± 1%     5.45s ± 2%     ~     (p=1.000 n=5+5)
RangeQuery/expr=changes(a_one[1d]),steps=1-12                                                                 609µs ± 0%     617µs ± 3%     ~     (p=0.190 n=4+5)
RangeQuery/expr=changes(a_one[1d]),steps=10-12                                                                801µs ± 6%     776µs ± 2%     ~     (p=0.310 n=5+5)
RangeQuery/expr=changes(a_one[1d]),steps=100-12                                                              2.65ms ±12%    2.35ms ± 1%     ~     (p=0.151 n=5+5)
RangeQuery/expr=changes(a_one[1d]),steps=1000-12                                                             19.3ms ± 8%    18.4ms ± 1%     ~     (p=0.056 n=5+5)
RangeQuery/expr=changes(a_ten[1d]),steps=1-12                                                                5.54ms ±10%    5.33ms ± 1%     ~     (p=0.548 n=5+5)
RangeQuery/expr=changes(a_ten[1d]),steps=10-12                                                               6.90ms ± 1%    6.90ms ± 2%     ~     (p=0.841 n=5+5)
RangeQuery/expr=changes(a_ten[1d]),steps=100-12                                                              23.3ms ± 4%    22.9ms ± 0%   -1.73%  (p=0.016 n=5+5)
RangeQuery/expr=changes(a_ten[1d]),steps=1000-12                                                              183ms ± 3%     182ms ± 1%     ~     (p=0.421 n=5+5)
RangeQuery/expr=changes(a_hundred[1d]),steps=1-12                                                            54.3ms ± 5%    53.6ms ± 0%     ~     (p=0.548 n=5+5)
RangeQuery/expr=changes(a_hundred[1d]),steps=10-12                                                           70.7ms ± 4%    69.6ms ± 0%   -1.53%  (p=0.032 n=5+5)
RangeQuery/expr=changes(a_hundred[1d]),steps=100-12                                                           237ms ± 7%     227ms ± 3%     ~     (p=0.056 n=5+5)
RangeQuery/expr=changes(a_hundred[1d]),steps=1000-12                                                          1.83s ± 2%     1.81s ± 0%   -1.36%  (p=0.032 n=5+5)
RangeQuery/expr=rate(a_one[1d]),steps=1-12                                                                    614µs ±11%     593µs ± 0%     ~     (p=0.151 n=5+5)
RangeQuery/expr=rate(a_one[1d]),steps=10-12                                                                   678µs ± 1%     686µs ± 2%     ~     (p=0.056 n=5+5)
RangeQuery/expr=rate(a_one[1d]),steps=100-12                                                                 1.60ms ± 9%    1.59ms ± 1%     ~     (p=0.310 n=5+5)
RangeQuery/expr=rate(a_one[1d]),steps=1000-12                                                                10.4ms ± 4%    10.7ms ± 1%     ~     (p=0.151 n=5+5)
RangeQuery/expr=rate(a_ten[1d]),steps=1-12                                                                   5.25ms ± 5%    5.13ms ± 0%   -2.32%  (p=0.008 n=5+5)
RangeQuery/expr=rate(a_ten[1d]),steps=10-12                                                                  5.95ms ± 1%    6.01ms ± 0%     ~     (p=0.190 n=5+4)
RangeQuery/expr=rate(a_ten[1d]),steps=100-12                                                                 14.9ms ± 0%    15.2ms ± 1%   +2.02%  (p=0.016 n=4+5)
RangeQuery/expr=rate(a_ten[1d]),steps=1000-12                                                                 103ms ± 8%     106ms ± 1%     ~     (p=0.151 n=5+5)
RangeQuery/expr=rate(a_hundred[1d]),steps=1-12                                                               51.8ms ± 1%    51.7ms ± 1%     ~     (p=0.690 n=5+5)
RangeQuery/expr=rate(a_hundred[1d]),steps=10-12                                                              60.1ms ± 1%    60.7ms ± 0%   +0.94%  (p=0.008 n=5+5)
RangeQuery/expr=rate(a_hundred[1d]),steps=100-12                                                              152ms ± 3%     153ms ± 3%     ~     (p=0.421 n=5+5)
RangeQuery/expr=rate(a_hundred[1d]),steps=1000-12                                                             1.04s ± 3%     1.04s ± 1%     ~     (p=0.690 n=5+5)
RangeQuery/expr=absent_over_time(a_one[1d]),steps=1-12                                                        588µs ± 0%     578µs ± 0%   -1.83%  (p=0.016 n=4+5)
RangeQuery/expr=absent_over_time(a_one[1d]),steps=10-12                                                       648µs ± 8%     596µs ± 0%   -8.05%  (p=0.008 n=5+5)
RangeQuery/expr=absent_over_time(a_one[1d]),steps=100-12                                                      849µs ± 9%     780µs ± 1%   -8.08%  (p=0.008 n=5+5)
RangeQuery/expr=absent_over_time(a_one[1d]),steps=1000-12                                                    2.63ms ± 5%    2.63ms ± 3%     ~     (p=0.841 n=5+5)
RangeQuery/expr=absent_over_time(a_ten[1d]),steps=1-12                                                       5.12ms ± 4%    4.97ms ± 1%   -2.88%  (p=0.032 n=5+5)
RangeQuery/expr=absent_over_time(a_ten[1d]),steps=10-12                                                      5.21ms ± 2%    5.17ms ± 1%     ~     (p=0.548 n=5+5)
RangeQuery/expr=absent_over_time(a_ten[1d]),steps=100-12                                                     6.98ms ± 1%    7.01ms ± 1%     ~     (p=0.421 n=5+5)
RangeQuery/expr=absent_over_time(a_ten[1d]),steps=1000-12                                                    26.3ms ± 5%    26.1ms ± 5%     ~     (p=1.000 n=5+5)
RangeQuery/expr=absent_over_time(a_hundred[1d]),steps=1-12                                                   52.7ms ± 6%    50.2ms ± 1%   -4.59%  (p=0.016 n=5+5)
RangeQuery/expr=absent_over_time(a_hundred[1d]),steps=10-12                                                  52.3ms ± 1%    52.2ms ± 0%     ~     (p=0.421 n=5+5)
RangeQuery/expr=absent_over_time(a_hundred[1d]),steps=100-12                                                 73.3ms ± 6%    71.0ms ± 1%     ~     (p=0.310 n=5+5)
RangeQuery/expr=absent_over_time(a_hundred[1d]),steps=1000-12                                                 265ms ± 7%     263ms ± 3%     ~     (p=0.841 n=5+5)
RangeQuery/expr=-a_one,steps=1-12                                                                            15.0µs ± 8%    13.2µs ± 1%  -11.67%  (p=0.008 n=5+5)
RangeQuery/expr=-a_one,steps=10-12                                                                           14.7µs ± 3%    13.6µs ± 1%   -7.35%  (p=0.029 n=4+4)
RangeQuery/expr=-a_one,steps=100-12                                                                          21.1µs ± 1%    19.7µs ± 1%   -6.72%  (p=0.008 n=5+5)
RangeQuery/expr=-a_one,steps=1000-12                                                                         73.6µs ±13%    63.8µs ± 0%  -13.36%  (p=0.008 n=5+5)
RangeQuery/expr=-a_ten,steps=1-12                                                                            60.8µs ±34%    52.4µs ± 0%  -13.87%  (p=0.008 n=5+5)
RangeQuery/expr=-a_ten,steps=10-12                                                                           64.9µs ± 9%    56.2µs ± 0%  -13.37%  (p=0.008 n=5+5)
RangeQuery/expr=-a_ten,steps=100-12                                                                           129µs ±11%     116µs ± 0%  -10.33%  (p=0.008 n=5+5)
RangeQuery/expr=-a_ten,steps=1000-12                                                                          617µs ± 5%     550µs ± 1%  -10.86%  (p=0.008 n=5+5)
RangeQuery/expr=-a_hundred,steps=1-12                                                                         474µs ± 0%     462µs ± 7%     ~     (p=0.151 n=5+5)
RangeQuery/expr=-a_hundred,steps=10-12                                                                        517µs ± 1%     494µs ± 1%   -4.43%  (p=0.008 n=5+5)
RangeQuery/expr=-a_hundred,steps=100-12                                                                      1.17ms ± 0%    1.10ms ± 3%   -6.14%  (p=0.008 n=5+5)
RangeQuery/expr=-a_hundred,steps=1000-12                                                                     5.94ms ± 1%    5.39ms ± 1%   -9.22%  (p=0.008 n=5+5)
RangeQuery/expr=a_one_-_b_one,steps=1-12                                                                     24.3µs ± 0%    23.1µs ± 2%   -4.78%  (p=0.016 n=4+5)
RangeQuery/expr=a_one_-_b_one,steps=10-12                                                                    31.3µs ± 5%    28.9µs ± 0%   -7.63%  (p=0.008 n=5+5)
RangeQuery/expr=a_one_-_b_one,steps=100-12                                                                   98.5µs ± 6%    91.5µs ± 0%   -7.11%  (p=0.008 n=5+5)
RangeQuery/expr=a_one_-_b_one,steps=1000-12                                                                   714µs ± 9%     678µs ± 0%   -4.91%  (p=0.008 n=5+5)
RangeQuery/expr=a_ten_-_b_ten,steps=1-12                                                                      127µs ± 5%     118µs ± 2%   -6.79%  (p=0.008 n=5+5)
RangeQuery/expr=a_ten_-_b_ten,steps=10-12                                                                     183µs ± 1%     177µs ± 0%   -3.30%  (p=0.008 n=5+5)
RangeQuery/expr=a_ten_-_b_ten,steps=100-12                                                                    812µs ± 1%     798µs ± 0%   -1.78%  (p=0.008 n=5+5)
RangeQuery/expr=a_ten_-_b_ten,steps=1000-12                                                                  6.79ms ± 2%    6.71ms ± 0%   -1.13%  (p=0.008 n=5+5)
RangeQuery/expr=a_hundred_-_b_hundred,steps=1-12                                                             1.11ms ± 1%    1.10ms ± 9%     ~     (p=0.190 n=4+5)
RangeQuery/expr=a_hundred_-_b_hundred,steps=10-12                                                            1.76ms ± 1%    1.74ms ± 1%   -1.34%  (p=0.008 n=5+5)
RangeQuery/expr=a_hundred_-_b_hundred,steps=100-12                                                           8.77ms ± 0%    8.69ms ± 1%   -0.99%  (p=0.008 n=5+5)
RangeQuery/expr=a_hundred_-_b_hundred,steps=1000-12                                                          76.2ms ± 0%    75.6ms ± 0%   -0.80%  (p=0.008 n=5+5)
RangeQuery/expr=a_one_-_b_one,steps=10000-12                                                                 6.91ms ± 0%    6.85ms ± 1%   -0.87%  (p=0.016 n=5+4)
RangeQuery/expr=a_ten_-_b_ten,steps=10000-12                                                                 69.6ms ± 1%    68.8ms ± 1%   -1.12%  (p=0.016 n=5+5)
RangeQuery/expr=a_hundred_-_b_hundred,steps=10000-12                                                          785ms ± 1%     772ms ± 2%   -1.69%  (p=0.016 n=5+5)
RangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1-12                                                    54.4µs ± 3%    52.7µs ± 3%     ~     (p=0.095 n=5+5)
RangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=10-12                                                   55.7µs ± 0%    54.4µs ± 0%   -2.34%  (p=0.016 n=4+5)
RangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=100-12                                                  85.2µs ± 0%    83.9µs ± 1%   -1.55%  (p=0.016 n=4+5)
RangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1000-12                                                  389µs ±19%     358µs ± 3%     ~     (p=0.151 n=5+5)
RangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=1-12                                                     137µs ±13%     123µs ± 1%  -10.19%  (p=0.008 n=5+5)
…