@roidelapluie 2021-03-11
Thanks for this. Did you consider adapting existing code so that it could handle both cases?