@pracucci 2021-03-12
> Thanks for this. Did you consider adapting existing code so that it could handle both usecases?

I tried a bit but looked a bit hacky the path I was following and some functions weren't working (eg. `Buffer()`). The proposed approach has some duplications (cons) but, in my opinion, a more easy to follow flow (pros). Considering this code is executed in a hot path (for PromQL engine) I would be more keen to err on the perf optimization side than perfectly deduplicated code, but I'm very open to ideas if we can remove dedup without compromising perfs.